ttt: portability (in commented code)
The change:

-# LANG=en_US.utf8 LC_ALL=en_US.utf8; export LANG LC_ALL
+# LANG=en_US.UTF-8 LC_ALL=en_US.UTF-8; export LANG LC_ALL

makes this code work in linux, osx (and probably elsewhere)

locale -a in linux does not print these, but it is shown as so
elsewhere (e.g. in /etc/locale.conf).

LANG=en_US.UTF-8 LC_ALL=en_US.UTF-8 perl

is good(ish) way to test when one uses working locale.

The code is commented out, but someone(tm) may copy and use it
somewhere (even in this file).